Fixed S0022-34, -37, -40
@rubywku 
S0022-34
In line 18 I changed deprel to mark.
In line 19 I changed deprel to advcl. 
For both changes, see bottom of "Grammar Notes" for explanation.
In line 23 I changed deprel to obl. Since the verbal noun is preceded by the article, this is behaving as a noun here, so it gets a nominal deprel.

S0022-37
Line 5 is spelled. It acts as the subject of the verb even though it is a conjugated preposition. I created the deprel obl:prep:arg for this (very rare) situation.
Sometimes, the non-nominative subject is a noun phrase cased-marked with a preposition. In such cases, the analysis is obl:arg for the noun.

S0022-40
Line 2:  Deprel is "dislocated" because the prepositional noun phrase is dislocated from its normal position after the main verb. If it were in fact after the main verb, its deprel would be xcomp:pred.
Line 4: The head of 7 because the adverb modifies the followed prepositional noun phrase.