Changed S0022–41, -46, -64
@rubywku 
S0022-41
Line 2: This is dislocated from its normal position after the main verb.

S0022-46
Line 5: This is obl:prep. Recall the general rule regarding prepositions that immediately precede relative clauses: they are obl:prep if the verb is not *ad-tá* and xcomp:pred:prep if the verb is *ad-tá*.
I added a new line for óen.

S0022-64
I added a new line for óen.